requirements.txt
asgiref==3.5.0
Django==3.2
django-dotenv==1.4.2
djangorestframework==3.13.1
pkg_resources==0.0.0
pytz==2021.3
sqlparse==0.4.2
gunicorn
requests
psycopg2-binary
django-storages
boto3